# backend/docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost:5000",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/playlists/{playlistId}" : {
      "put" : {
        "tags" : [ "playlist-controller" ],
        "summary" : "Edit a playlist's details",
        "description" : "Partial update of a playlist the caller owns or can edit (accepted EDITOR). Only the owner can change `isPublic`. Every multipart field is optional; a missing or\nblank field leaves that value unchanged.\n- `playlistName`: at most 36 chars, trimmed.\n- `isPublic`: `true` or `false`.\n- `playlistCover`: same rules as on create. Replaces the old cover.\nReturns the playlist's id, name and cover URL after the update.\n",
        "operationId" : "editPlaylist",
        "parameters" : [ {
          "name" : "playlistId",
          "in" : "path",
          "description" : "Public id of the playlist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "p_7c2dK1"
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlaylistDetailsUpdate"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CorePlaylist"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad field, or unreadable or too small cover (codes INVALID_FIELD, INVALID_IMAGE, IMAGE_TOO_SMALL)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "403" : {
            "description" : "Editor tried to change `isPublic` (code PLAYLIST_ACTION_FORBIDDEN); or Account has not finished onboarding (code ONBOARDING_INCOMPLETE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "404" : {
            "description" : "Playlist not found, or caller can't edit it (code PLAYLIST_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "413" : {
            "description" : "Cover over 5MB or 25 megapixels (code IMAGE_TOO_LARGE), or file over 25MB (code REQUEST_FAILED)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "415" : {
            "description" : "Cover is not a JPEG or PNG (code UNSUPPORTED_IMAGE_TYPE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid token (code INVALID_TOKEN)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 20 requests per minute per IP, shared by all upload endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/api/playlists/{playlistId}/save" : {
      "put" : {
        "tags" : [ "playlist-controller" ],
        "summary" : "Save a playlist",
        "description" : "Adds another user's public playlist to the caller's saved playlists. Saving a playlist\nthat is already saved is a no-op and still returns 200.\nPrivate playlists return 404, the same as a missing playlist.\nReturns the playlist's id, name and cover URL, and `isSaved` (always true).\n",
        "operationId" : "savePlaylist",
        "parameters" : [ {
          "name" : "playlistId",
          "in" : "path",
          "description" : "Public id of the playlist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "p_7c2dK1"
        } ],
        "responses" : {
          "200" : {
            "description" : "Playlist saved (or already saved)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SavedPlaylistResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller owns the playlist or is a member of it (code PLAYLIST_ACTION_FORBIDDEN); or Account has not finished onboarding (code ONBOARDING_INCOMPLETE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "404" : {
            "description" : "Playlist not found, or private (code PLAYLIST_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid token (code INVALID_TOKEN)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 120 requests per minute per IP, shared by all other endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "playlist-controller" ],
        "summary" : "Unsave a playlist",
        "description" : "Removes a playlist from the caller's saved playlists. Unsaving a playlist the caller\nhas not saved returns 404, so a repeated call is not idempotent.\nReturns the playlist's id, name and cover URL, and `isSaved` (always false).\n",
        "operationId" : "unsavePlaylist",
        "parameters" : [ {
          "name" : "playlistId",
          "in" : "path",
          "description" : "Public id of the playlist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "p_7c2dK1"
        } ],
        "responses" : {
          "200" : {
            "description" : "Playlist unsaved",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SavedPlaylistResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Caller owns the playlist (code PLAYLIST_ACTION_FORBIDDEN); or Account has not finished onboarding (code ONBOARDING_INCOMPLETE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "404" : {
            "description" : "Playlist missing, or not saved by the caller (code SAVED_PLAYLIST_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid token (code INVALID_TOKEN)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 120 requests per minute per IP, shared by all other endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/api/playlists/{playlistId}/pieces" : {
      "post" : {
        "tags" : [ "playlist-controller" ],
        "summary" : "Add music pieces to a playlist",
        "description" : "Owner or accepted editor only. Appends the pieces to the end of the playlist, in the order the server finds them.\nA piece is skipped (counted in `failedCount`) when its id is unknown, it is already in the playlist, or it is private.\nA private piece is only accepted from its uploader, and only into a non-public playlist or one the uploader owns.\nSkipped pieces do not fail the request.\n",
        "operationId" : "addPieceToPlaylist",
        "parameters" : [ {
          "name" : "playlistId",
          "in" : "path",
          "description" : "Public id of the playlist",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "p_7c2dK1"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlaylistMusicPieceAddition"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK, even if every piece was skipped",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlaylistMusicPieceAdditionResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "`musicPieceIds` missing, empty or over 100 (code INVALID_FIELD)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "404" : {
            "description" : "Playlist not found, or caller can't edit it (code PLAYLIST_NOT_FOUND)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid token (code INVALID_TOKEN)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "403" : {
            "description" : "Account has not finished onboarding (code ONBOARDING_INCOMPLETE)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "429" : {
            "description" : "Too many requests from this IP (code RATE_LIMITED). Rate limit: 120 requests per minute per IP, shared by all other endpoints; short bursts are fine. Wait `Retry-After` seconds before retrying.",
            "headers" : {
              "Retry-After" : {
                "description" : "Seconds until the next request is allowed",
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          },
          "500" : {
            "description" : "Unexpected server error (code INTERNAL_ERROR)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuxServerError"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "PlaylistDetailsUpdate" : {
        "type" : "object",
        "description" : "Multipart form for editing a playlist; a missing field leaves that value unchanged",
        "properties" : {
          "playlistName" : {
            "type" : "string",
            "description" : "New name, at most 36 chars. Omit or leave blank to keep the current name",
            "example" : "My Playlist",
            "maxLength" : 36,
            "minLength" : 0
          },
          "isPublic" : {
            "type" : "boolean",
            "description" : "New visibility. Owner only; omit to keep the current value",
            "example" : true
          },
          "playlistCover" : {
            "type" : "string",
            "format" : "binary",
            "description" : "New cover: JPEG or PNG, at most 5MB and 25 megapixels, at least 256x256. Omit to keep the current cover"
          }
        }
      },
      "CorePlaylist" : {
        "type" : "object",
        "description" : "The basic identity of a playlist",
        "properties" : {
          "playlistId" : {
            "type" : "string",
            "description" : "Public id of the playlist",
            "example" : "p_7c2dK1"
          },
          "playlistName" : {
            "type" : "string",
            "description" : "Display name, 1-36 chars",
            "example" : "Late Night Drive"
          },
          "playlistCoverUrl" : {
            "type" : "string",
            "description" : "Square WebP cover, at most 1024x1024",
            "example" : "https://aux.justinabruinsma.com/playlist-cover/3f2b8c1e-8d4a-4c6e-9a51-2f0d7b9e6c11.webp"
          }
        }
      },
...
      "AuxServerError" : {
        "description" : "Body of every 4xx and 5xx response",
        "properties" : {
          "errorDetails" : {
            "$ref" : "#/components/schemas/ErrorDetails"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Token from /api/auth/login or /api/auth/register",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}